Added more config possibilities, switched to sending timestams as json text and then parse back, because json supports only 32 bit ints.

+++ src/messages/MessageUtils.hpp
@@ -1,4 +1,7 @@
 #pragma once
 
+#include <cstdint>
+#include <memory>
+
 int generate_unique_id();
-int generate_unique_timestamp();
+int64_t generate_unique_timestamp();
